Split the panels up so that it is easier to read and more modular.
